Encryption: add decryption capability
The 'derypt_incassoos_data' capability is only assigned to the Collector role. The capability provides access to enabling and disabling the encryption in Incassoos.

Encryption of data in itself should be available for all users as it is a mechanic that is forced on the registered data values. Decryption of data should only be available for the appointed roles, however for now the decryption key as password works as well.