Encryption: introduce decryption key handlers
Adds get/set/delete functions for the decryption key for usage within a single page request.
The decryption key is treated as a password-like value in that the value is not stored in any place, but it is not hashed because the original value is needed for decrypting the encrypted values.