Export: optionally require decryption key per export type
Export types that use encrypted data in their exports can be registered as requiring the decryption key before exporting.
When the decryption key is required, its presence will be checked before executing the export. Changes in the UI for providing the decryption key will be added in a following commit.