Export: move handling of file downloads to admin-post.php
With the introduction of the Consumers export type, not all file downloads have a post context. To make the handling of file downloads post-agnostic, let's use the hooks provided in admin-post.php instead of post.php.
Consequently the file download links require the `post` query argument to indicate when the donwload concerns a particular post.

Additionally, the export details will not be deleted on first download so that re-downloads are possible untill the export details transient is expired.