Introduce the 'Unknown user' consumer type
The Unknown user is a fallback mechanism for identifying consumer data of users that could not be found by the users query. Without this, consumer data that relates to deleted or otherwise missing user IDs, the data will be omitted when consumer data is requested per consumer. This is the case for example in post exports, like Collection consumers.

The new consumer type can act both as a global container for all associated missing users as well as a per-user key when combined with the user ID. The `<consumer-type-id>:ID` type syntax will be interpreted as the Unknown user item of a specific user so that individual consumer data is still available to identify.

This change also introduces consumer types for Activity participants, named participant types. For now, prices for participant types can be shown for closed activities, but they are not available for registering activity prices yet.

The logic for the VGSR SFC export type is adjusted to account for the new Activity participant types.