Admin: add taxonomy filters to the plugin post list tables
The list tables can now be filtered by use of the new taxonomy dropdowns.

To make this work, new request query vars are added to the set of public request query vars in WP. For now only the plugin taxonomy names are available query vars. Their respective values are expected to be numeric term ids.