Query: redefine what empty assets are for the implementation of the 'incassoos_empty' parameter.
Previously empty assets where those with a total value of '0'. However, since the possibility
of negative values, the total value of an asset might be '0' while actually having valid registrations.

The new definition of empty assets resulted in the following logic:
- Activities are empty when they have no participants
- Occasions are empty when they have no orders

This commit implements these new definitions when parsing query vars for Activities and when filtering query clauses for Occasions.