App: restructure sections for medium sized screens
Introduces the new Resize Service which triggers event listeners for specified breakpoints so that
they are handled in a central place. Definition of the breakpoints still happens in the Sections component.

In the medium screen size, the new setup has the consumers section as always active, instead of the order panel.
This should reduce clicks and increases focus on entering new orders, without the need to switch sections.