REST: declare additional fields available for the application
To prevent having large objects in the application due to unneeded object properties, we need to limit the fields that are loaded from the endpoint. A solution is to next to the app's base fields, only process specifically declared additional fields in the application. The declared fields are listed in the `_applicationFields` field in the endpoint. This list is dynamically derived from the additional rest fields that are added to the endpoint through `register_rest_field()`.