Activity: fix error when duplicating an item with an activity date
When duplicating an Activity, the validation on the activity date failed because the value was provided from the raw metadata SQL date format.
The fix consists of moving the filter for the duplication post args to include the metadata args and defining an additional filter for activity items that modify the date format prior to sending the post args to `wp_insert_post()` where the validation is handled.