Introduce custom Consumer Types
Using a new taxonomy users can create their own Consumer Types next to the built-in ones. Getters and query functions are updated to accomodate for Consumer Types based on terms. Filters are in place to prevent creating terms that are duplicates of built-in Consumer Types.

Custom Consumer Types are managed in the common taxonomy page which is accessed through the Consumers admin page.