Change hidden objects to archived objects
This applies to Product Categories (taxonomy) and Consumers.

Using the concept of archiving, it should be easier to distinguish between
hiding an object in the interface versus changing a setting on an object so
it is made unavailable. Both for the user and for the developer.

This change includes renaming object functions, REST endpoints, interface
elements and updating metadata in the database.

For Consumers the default screen now shows public items only. Archived items
are shown after toggling the 'Show archived' button.

This change is a prelude to more archiving of objects.